+++ freekick/__init__.py
@@ -8,8 +8,6 @@ from .utils.freekick_logging import _logger
 
 ROOT_DIR = Path(__file__).parent.parent
 DATA_DIR = ROOT_DIR / "data"
-
-# TODO LEARNER_MODEL_LOCATION should be tmp_dir/azure blob, gcs, s3 etc..
 ESTIMATOR_LOCATION = DATA_DIR / "learner_model"
 
 # Ensure we are using the min_supported_platform
@@ -25,5 +23,3 @@ if not _PLATFORM_CHECK:
             f" Python version is {min_supported_platform}"
         )
     _PLATFORM_CHECK = True
-
-# TODO: Replace all print statements with log statements