+++ freekick/api/__init__.py
@@ -1,6 +1,3 @@
-# TODO: API requests also take arg which specify which model/learner is wants
-# to use. Will be needed once we have more than one model at a time. This will
-# make it easier to switch out models at any time.
 from flask_restx import Api
 
 from freekick import __version__
@@ -11,7 +8,11 @@ from .match import match_ns
 from .season import season_ns
 
 freekick_api = Api(
-    title="Freekick", version=str(__version__), prefix="/api", validate=True
+    title="Freekick",
+    version=str(__version__),
+    prefix="/api",
+    validate=True,
+    doc="/swagger",
 )
 
 freekick_api.add_namespace(match_ns)