+++ freekick/api/match.py
@@ -1,7 +1,10 @@
 """API that calls the match_predictor service"""
 
+from datetime import time
+
 from flask_restx import Namespace, Resource, fields, reqparse
 
+from freekick.datastore.util import League
 from freekick.service import predict_match
 
 match_ns = Namespace("match", description="Single match operations.")
@@ -25,6 +28,8 @@ post_parser.add_argument("league", type=str, required=True, help="League code")
 post_parser.add_argument(
     "attendance", type=float, help="Approximate attendance"
 )
+post_parser.add_argument("match_time", type=str, help="Match Time")
+post_parser.add_argument("match_date", type=str, help="Match Date")
 
 
 @match_ns.route("/")
@@ -34,14 +39,12 @@ class MatchApi(Resource):
     @match_ns.expect(post_parser)
     def post(self):
         args = post_parser.parse_args(strict=True)
-        # TODO: Also get date and time from front ends request and pass to
-        # predict_match
         match_dto = predict_match(
-            # TODO should inherit and use League enum here or have
-            # predict_match create the enum?
-            league=args["league"],
+            league=League[args["league"].upper()],
             home_team=args["home_team"].replace(" ", "-"),
             away_team=args["away_team"].replace(" ", "-"),
             attendance=args["attendance"],
+            time=args["match_time"],
+            match_date=args["match_date"] or None,
         )
         return match_dto