+++ freekick/learners/regression.py
@@ -1,103 +1,31 @@
 """Model for various logistic models."""
 
-import pickle
-
-import pandas as pd
 from dask_ml.linear_model import LogisticRegression as dask_LR
 from sklearn.linear_model import LogisticRegression as sklearn_LR
 
-from freekick import ESTIMATOR_LOCATION, _logger
-from freekick.datastore.util import Backend
+from freekick import _logger
+from freekick.datastore.util import Backend, League
 
-# TODO: Fix or eliminate SoccerLogisticModel to inherit from .classification
-# BaseEstimator
+from .classification import BaseClassifier
 
 
-class SoccerLogisticModel:
+class SoccerLogisticModel(BaseClassifier):
     def __init__(
-        self,
-        league: str,
-        X: pd.DataFrame,
-        y: pd.Series,
-        backend: Backend = Backend.PANDAS,
+        self, league: League, backend: Backend = Backend.PANDAS
     ) -> None:
-        self.league = league
-        self.model = None
-        self.y = y
-        self.X = X
         self.backend = backend
+        super().__init__(league)
 
-        self.initialize_backend()
+    def init_model(self):
+        match self.backend:
+            case Backend.PANDAS:
+                learner = sklearn_LR
+            case Backend.DASK:
+                learner = dask_LR
+            case _:
+                raise ValueError(
+                    f"{self.__class__} does not support backend {self.backend}"
+                )
 
-    def initialize_backend(self):
-        self.logistic_regression = (
-            dask_LR if self.backend == Backend.DASK else sklearn_LR
-        )
-        _logger.info(f"Selected Backend: {self.logistic_regression}")
-
-    def fit(self):
-        model = self.logistic_regression(
-            penalty="l2", fit_intercept=False, C=1
-        )
-        self.model = model.fit(self.X, self.y)
-
-    def check_fit(self):
-        if not self.model:
-            raise NotImplementedError(
-                "Model not trained: Please train/fit the model first using"
-                "SoccerLogisticModel.fit()."
-            )
-
-    def get_coeff(self):
-        # TODO: .coef_ does not exists for DecisionTreeClassifier
-        # get the coefficients of three logistic models for each team.
-        # self.check_fit()
-
-        # coeffs = pd.DataFrame(
-        #     self.model.coef_, index=self.model.classes_, columns=self.X.columns
-        # ).T
-        # coeffs = coeffs.rename(
-        #     columns={-1: "away_win", 0: "draw", 1: "home_win"}
-        # )
-
-        # return coeffs
-        return None
-
-    def predict(self, pred_data):
-        """Predict the match winner"""
-        self.check_fit()
-
-        pred = self.model.predict(pred_data)
-
-        df_pred = pd.DataFrame(
-            pred, index=pred_data.index, columns=["Prediction"]
-        )
-        return df_pred
-
-    def predict_probability(self, X):
-        """Predict the probabilities of draw or win for each team"""
-        self.check_fit()
-
-        df = pd.DataFrame(
-            self.model.predict_proba(X),
-            columns=self.model.classes_,
-            index=X.index,
-        )
-        df = df.rename(columns={-1: "away_win", 0: "draw", 1: "home_win"})
-
-        return df
-
-    def persist_model(self):
-        """Persists a model as Pickle file on disk in models folder.
-        Overwrite if file already exists.
-
-        Parameters
-        ----------
-        name : str, optional
-            Name of the pickle file, by default 'soccer_model'
-        """
-        self.check_fit()
-        file_path = ESTIMATOR_LOCATION / f"{self.name}.pkl"
-        with open(file_path, "wb") as mod_file:
-            pickle.dump(self.model, mod_file)
-        print(f"Model '{self.league}' serialized to {file_path}.")
+        _logger.info(f"Selected Backend: {self.backend}")
+        return learner(penalty="l2", fit_intercept=False, C=1)