+++ freekick/service/match_day_predictor.py
@@ -2,7 +2,7 @@ from .util import MatchDTO
 
 
 def predict_match_day(league: str):
-    # TODO: Complete function
+    raise NotImplementedError
     return [
         MatchDTO(
             home_team="Team1", away_team="Team2", predicted_winner="TEAM1"